+++ craft_application/application.py
@@ -34,6 +34,7 @@ from typing import TYPE_CHECKING, Any, Literal, cast, final
 
 import craft_cli
 import craft_parts
+import craft_platforms
 import craft_providers
 from craft_parts.plugins.plugins import PluginType
 from platformdirs import user_cache_path
@@ -616,8 +617,10 @@ class Application:
             craft_cli.emit.ended_ok()
             return_code = os.EX_USAGE
         except KeyboardInterrupt as err:
-            self._emit_error(craft_cli.CraftError("Interrupted."), cause=err)
             return_code = 128 + signal.SIGINT
+            self._emit_error(
+                craft_cli.CraftError("Interrupted.", retcode=return_code), cause=err
+            )
         except craft_cli.CraftError as err:
             self._emit_error(err)
             return_code = err.retcode
@@ -635,14 +638,42 @@ class Application:
                 cause=err,
             )
             return_code = 1
-        except Exception as err:
+        except craft_platforms.CraftPlatformsError as err:
             self._emit_error(
-                craft_cli.CraftError(f"{self.app.name} internal error: {err!r}"),
+                craft_cli.CraftError(
+                    err.args[0],
+                    details=err.details,
+                    resolution=err.resolution,
+                    reportable=err.reportable,
+                    docs_url=err.docs_url,
+                    doc_slug=err.doc_slug,
+                    logpath_report=err.logpath_report,
+                    retcode=err.retcode,
+                ),
                 cause=err,
             )
+            return_code = err.retcode
+        except Exception as err:
+            if isinstance(err, craft_platforms.CraftError):
+                transformed = craft_cli.CraftError(
+                    err.args[0],
+                    details=err.details,
+                    resolution=err.resolution,
+                    docs_url=getattr(err, "docs_url", None),
+                    doc_slug=getattr(err, "doc_slug", None),
+                    logpath_report=getattr(err, "logpath_report", True),
+                    reportable=getattr(err, "reportable", True),
+                    retcode=getattr(err, "retcode", 1),
+                )
+                return_code = transformed.retcode
+            else:
+                transformed = craft_cli.CraftError(
+                    f"{self.app.name} internal error: {err!r}"
+                )
+                return_code = os.EX_SOFTWARE
+            self._emit_error(transformed, cause=err)
             if self.services.get("config").get("debug"):
                 raise
-            return_code = os.EX_SOFTWARE
         else:
             craft_cli.emit.ended_ok()