+++ docs/reference/changelog.rst
@@ -11,6 +11,7 @@ Commands
 ========
 
 - ``test`` raises a clear error message if ``spread.yaml`` is missing.
+- The warning that the ``test`` command is experimental is only displayed once.
 
 Services
 ========